// src/api/generated/models/gameSessionStatus.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * TDD Game API
 * OpenAPI spec version: v1
 */

export type GameSessionStatus = typeof GameSessionStatus[keyof typeof GameSessionStatus];


export const GameSessionStatus = {
  lobby: 'lobby',
  started: 'started',
  completed: 'completed',
} as const;